split configuration-helpers into aas and non-aas helpers
